Move e2e tests
