+++ bookmarks/tests_e2e/e2e_test_bookmark_details_modal.py
@@ -2,7 +2,7 @@ from django.test import override_settings
 from django.urls import reverse
 from playwright.sync_api import sync_playwright, expect
 
-from bookmarks.e2e.helpers import LinkdingE2ETestCase
+from bookmarks.tests_e2e.helpers import LinkdingE2ETestCase
 from bookmarks.models import Bookmark
