+++ bookmarks/tests_e2e/e2e_test_bookmark_item.py
@@ -3,7 +3,7 @@ from unittest import skip
 from django.urls import reverse
 from playwright.sync_api import sync_playwright, expect
 
-from bookmarks.e2e.helpers import LinkdingE2ETestCase
+from bookmarks.tests_e2e.helpers import LinkdingE2ETestCase
 
 
 class BookmarkItemE2ETestCase(LinkdingE2ETestCase):
