+++ bookmarks/tests_e2e/e2e_test_edit_bookmark_form.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 from django.urls import reverse
 from playwright.sync_api import sync_playwright, expect
 
-from bookmarks.e2e.helpers import LinkdingE2ETestCase
+from bookmarks.tests_e2e.helpers import LinkdingE2ETestCase
 from bookmarks.services import website_loader
 
 mock_website_metadata = website_loader.WebsiteMetadata(
