+++ bookmarks/tests_e2e/e2e_test_filter_drawer.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 from playwright.sync_api import sync_playwright, expect
 
-from bookmarks.e2e.helpers import LinkdingE2ETestCase
+from bookmarks.tests_e2e.helpers import LinkdingE2ETestCase
 
 
 class FilterDrawerE2ETestCase(LinkdingE2ETestCase):
