+++ bookmarks/tests_e2e/e2e_test_global_shortcuts.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 from playwright.sync_api import sync_playwright, expect
 
-from bookmarks.e2e.helpers import LinkdingE2ETestCase
+from bookmarks.tests_e2e.helpers import LinkdingE2ETestCase
 
 
 class GlobalShortcutsE2ETestCase(LinkdingE2ETestCase):
